requirements.txt
pyzmq>=25.1.0
quickfix>=1.15.1
simplefix>=1.0.15
pydantic>=2.5.0
python-dateutil>=2.8.2
pandas>=2.1.0
numpy>=1.24.0
loguru>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0
dataclasses-json>=0.6.0
typing-extensions>=4.8.0
